+++ contracts/0.8.9/oracle/ValidatorsExitBusOracle.sol
@@ -10,7 +10,7 @@ import {BaseOracle} from "./BaseOracle.sol";
 import {ValidatorsExitBus} from "./ValidatorsExitBus.sol";
 
 interface IOracleReportSanityChecker {
-    function checkExitBusOracleReport(uint32 _maxBalanceExitRequestedPerReportInEth) external view;
+    function checkExitBusOracleReport(uint16 _maxBalanceExitRequestedPerReportInEth) external view;
 }
 
 contract ValidatorsExitBusOracle is BaseOracle, ValidatorsExitBus {
@@ -242,9 +242,9 @@ contract ValidatorsExitBusOracle is BaseOracle, ValidatorsExitBus {
             revert UnexpectedRequestsDataLength();
         }
 
-        // Calculate total balance of validators being exited in ETH (uint32)
+        // Calculate total balance of validators being exited in ETH (uint16)
         // Module 1 (curated) uses 32 ETH, other modules use 2048 ETH per validator
-        uint32 totalExitBalanceEth = _calculateTotalExitBalanceEth(data.data, data.dataFormat);
+        uint16 totalExitBalanceEth = _calculateTotalExitBalanceEth(data.data, data.dataFormat);
 
         // Pass ETH value directly to sanity checker
         IOracleReportSanityChecker(LOCATOR.oracleReportSanityChecker()).checkExitBusOracleReport(totalExitBalanceEth);
