+++ test/0.8.9/contracts/ValidatorsExitBus__Harness.sol
@@ -72,7 +72,7 @@ contract ValidatorsExitBus__Harness is ValidatorsExitBusOracle, ITimeProvider {
     }
 
     // Expose internal function for unit testing
-    function calculateTotalExitBalanceEth(bytes calldata data, uint256 dataFormat) external view returns (uint32) {
+    function calculateTotalExitBalanceEth(bytes calldata data, uint256 dataFormat) external view returns (uint16) {
         return _calculateTotalExitBalanceEth(data, dataFormat);
     }
 }
